requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
djangorestframework==3.11.0
idna==2.9
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
unittest-xml-reporting==3.0.2
urllib3==1.25.8
django-url-filter